+++ PointsBot.py
@@ -1,7 +1,9 @@
 """Entry point used for either runnning or freezing the bot."""
 import pointsbot
 
-try:
-    pointsbot.run()
-except KeyboardInterrupt as e:
-    print('\nShutting down...\n')
+while True:
+    try:
+        pointsbot.run()
+    except KeyboardInterrupt as e:
+        print('\nShutting down...\n')
+        break