+++ pointsbot/bot.py
@@ -11,11 +11,11 @@ from . import config, database, level, reply
 
 ### Globals ###
 
-USER_AGENT = 'PointsBot (by u/GlipGlorp7)'
+USER_AGENT = 'PointsBot (by u/JustSomeStuffIDid)'
 
-# The pattern that determines whether a post is marked as solved
-SOLVED_PATTERN = re.compile('![Hh]elped')
-MOD_SOLVED_PATTERN = re.compile('/[Hh]elped')
+# The pattern that determines whether a post is marked as helpful
+THANKED_PATTERN = re.compile(r'\b([Tt]hanks)(?!\s+to)\b|([Tt]hank [Yy]ou)|\b([Tt]y)\b|\b([Tt]hx)\b')
+MOD_THANKED_PATTERN = re.compile('/[Hh]elped')
 MOD_REMOVE_PATTERN = re.compile('/[Rr]emove[Pp]oint')
 
 
@@ -70,7 +70,7 @@ def run():
 
 
 def monitor_comments(reddit, subreddit, db, levels, cfg):
-    """Monitor new comments in the subreddit, looking for confirmed solutions."""
+    """Monitor new comments in the subreddit, looking for confirmed awards."""
     # Passing pause_after=0 will bypass the internal exponential delay, but have
     # to check if any comments are returned after each query
     for comm in subreddit.stream.comments(skip_existing=True, pause_after=0):
@@ -87,13 +87,13 @@ def monitor_comments(reddit, subreddit, db, levels, cfg):
         logging.debug('Comment author: "%s"', comm.author.name)
         logging.debug('Comment text: "%s"', comm.body)
 
-        mark_as_solved, remove_point, is_mod_command = marks_as_solved(comm)
-        if mark_as_solved:
-            logging.info('Comment marks issue as solved')
+        award_point, remove_point, is_mod_command = awards_point(comm)
+        if award_point:
+            logging.info('Comment marks awards point')
         elif remove_point:
             logging.info('Comment removes point')
         else:
-            # Skip this "!solved" comment
+            # Skip this comment
             logging.info('Comment does not have a valid command')
             continue
 
@@ -102,33 +102,33 @@ def monitor_comments(reddit, subreddit, db, levels, cfg):
         elif is_valid_tag(comm, cfg.tags):
             logging.info('Comment has a valid tag')
 
-        solver, solution_comment = find_solver_and_comment(comm)
-        solver_has_already_solved = db.has_already_solved_once(solution_comment.submission, solver)
-        if not remove_point and solver_has_already_solved:
-            logging.info('User "%s" has already solved this submission once', solver.name)
+        helper, helper_comment = find_helper_and_comment(comm)
+        helper_has_already_helped = db.has_already_helped_once(helper_comment.submission, helper)
+        if not remove_point and helper_has_already_helped:
+            logging.info('User "%s" has already been awarded for this submission once', helper.name)
             logging.info('No additional points awarded')
             continue
 
         if remove_point:
-            db.soft_remove_point_for_solution(comm.submission, solver, comm.author, comm)
-            logging.info('Removed point for user "%s"', solver.name)
+            db.soft_remove_point_for_answer(comm.submission, helper, comm.author, comm)
+            logging.info('Removed point for user "%s"', helper.name)
         else:
-            logging.info('Submission solved')
-            logging.debug('Solution comment:')
-            logging.debug('Author: %s', solution_comment.author.name)
-            logging.debug('Body:   %s', solution_comment.body)
-            db.add_point_for_solution(comm.submission, solver, solution_comment, comm.author, comm)
-            logging.info('Added point for user "%s"', solver.name)
+            logging.info('Post awarded')
+            logging.debug('Helper comment:')
+            logging.debug('Author: %s', helper_comment.author.name)
+            logging.debug('Body:   %s', helper_comment.body)
+            db.add_point_for_answer(comm.submission, helper, helper_comment, comm.author, comm)
+            logging.info('Added point for user "%s"', helper.name)
 
-        points = db.get_points(solver)
-        logging.info('Total points for user "%s": %d', solver.name, points)
+        points = db.get_points(helper)
+        logging.info('Total points for user "%s": %d', helper.name, points)
         if points > 0:
             level_info = level.user_level_info(points, levels)
         else:
             level_info = None
 
-        # Reply to the comment marking the submission as solved
-        reply_body = reply.make(solver,
+        # Reply to the comment thanking the helper
+        reply_body = reply.make(helper,
                                 points,
                                 level_info,
                                 feedback_url=cfg.feedback_url,
@@ -141,11 +141,11 @@ def monitor_comments(reddit, subreddit, db, levels, cfg):
         except praw.exceptions.APIException as e:
             logging.error('Unable to reply to comment: %s', e)
             if remove_point:
-                db.add_back_point_for_solution(comm.submission, solver)
-                logging.error('Re-added point that was just removed from user "%s"', solver.name)
+                db.add_back_point_for_answer(comm.submission, helper)
+                logging.error('Re-added point that was just removed from user "%s"', helper.name)
             else:
-                db.remove_point_and_delete_solution(comm.submission, solver)
-                logging.error('Removed point that was just awarded to user "%s"', solver.name)
+                db.remove_point_and_delete_answer(comm.submission, helper)
+                logging.error('Removed point that was just awarded to user "%s"', helper.name)
             logging.error('Skipping comment')
             continue
 
@@ -153,88 +153,94 @@ def monitor_comments(reddit, subreddit, db, levels, cfg):
         if level_info and level_info.current and level_info.current.points == points:
             lvl = level_info.current
             logging.info('User reached level: %s', lvl.name)
-            if not subreddit.moderator(redditor=solver):
+            if not subreddit.moderator(redditor=helper):
                 logging.info('User is not mod; setting flair')
                 logging.info('Flair text: %s', lvl.name)
                 logging.info('Flair template ID: %s', lvl.flair_template_id)
-                subreddit.flair.set(solver,
+                subreddit.flair.set(helper,
                                     text=lvl.name,
                                     flair_template_id=lvl.flair_template_id)
             else:
-                logging.info('Solver is mod; don\'t alter flair')
+                logging.info('helper is mod; don\'t alter flair')
 
 
 ### Reddit Comment Functions ###
 
-SolutionResponseRule = namedtuple('SolutionResponseRule',
+AnswerResponseRule = namedtuple('AnswerResponseRule',
                                   'description success_msg failure_msg check')
 
 OP_RESPONSE_RULES = [
-    SolutionResponseRule(
-        'user "solved" pattern',
-        'Comment contains user "solved" pattern',
-        'Comment does not contain user "solved" pattern',
-        lambda c: SOLVED_PATTERN.search(c.body),
+    AnswerResponseRule(
+        'user "thanked" pattern',
+        '[PASS] Comment contains user "thanked" pattern',
+        '[FAIL] Comment does not contain user "thanked" pattern',
+        lambda c: THANKED_PATTERN.search(c.body),
     ),
-    SolutionResponseRule(
+    AnswerResponseRule(
         'is a reply (not top-level)',
-        'Comment is a reply to another comment',
-        'Comment is a top-level comment',
+        '[PASS] Comment is a reply to another comment',
+        '[FAIL] Comment is a top-level comment',
         lambda c: not c.is_root,
     ),
-    SolutionResponseRule(
+    AnswerResponseRule(
         'author is OP',
-        'Comment author is submission OP',
-        'Comment author is not submission OP',
+        '[PASS] Comment author is thread OP',
+        '[FAIL] Comment author is not thread OP',
         lambda c: c.is_submitter,
     ),
-    SolutionResponseRule(
-        "OP can't solve own problem",
-        'Submission OP is different from solution author',
-        'Submission OP is marking own comment as solution',
+    AnswerResponseRule(
+        "OP of thread can't be awarded",
+        '[PASS] Thanked comment author is different from thread author',
+        "[FAIL] Thread author is thanking themselves",
         lambda c: not c.is_root and not c.parent().is_submitter,
+    ),
+    AnswerResponseRule(
+        "Mod can't be awarded",
+        '[PASS] Thanked comment author is not mod',
+        "[FAIL] Thanked comment author is a mod",
+        lambda c: not is_mod_comment(c.parent()),
     ),
 ]
 
 MOD_RESPONSE_RULES = [
-    SolutionResponseRule(
-        'contains mod "solved" pattern',
-        'Comment contains mod "solved" pattern',
-        'Comment does not contain mod "solved" pattern',
-        lambda c: MOD_SOLVED_PATTERN.search(c.body),
+    AnswerResponseRule(
+        'contains mod "thanked" pattern',
+        '[PASS] Comment contains mod "thanked" pattern',
+        '[FAIL] Comment does not contain mod "thanked" pattern',
+        lambda c: MOD_THANKED_PATTERN.search(c.body),
     ),
-    SolutionResponseRule(
+    AnswerResponseRule(
         'is a reply (not top-level)',
-        'Comment is a reply to another comment',
-        'Comment is a top-level comment',
+        '[PASS] Comment is a reply to another comment',
+        '[FAIL] Comment is a top-level comment',
         lambda c: not c.is_root,
     ),
-    SolutionResponseRule(
+    AnswerResponseRule(
         'author is mod',
-        'Comment author is a mod',
-        'Comment author is not a mod',
+        '[PASS] Comment author is a mod',
+        '[FAIL] Comment author is not a mod',
         # TODO Initialize rules in a function so that they can include other functions
         lambda c: is_mod_comment(c),
     ),
 ]
 
 MOD_REMOVE_RULES = [
-    SolutionResponseRule(
+    AnswerResponseRule(
         'contains mod "removepoint" pattern',
-        'Comment contains mod "removepoint" pattern',
-        'Comment does not contain mod "removepoint" pattern',
+        '[PASS] Comment contains mod "removepoint" pattern',
+        '[FAIL] Comment does not contain mod "removepoint" pattern',
         lambda c: MOD_REMOVE_PATTERN.search(c.body),
     ),
-    SolutionResponseRule(
+    AnswerResponseRule(
         'is a reply (not top-level)',
-        'Comment is a reply to another comment',
-        'Comment is a top-level comment',
+        '[PASS] Comment is a reply to another comment',
+        '[FAIL] Comment is a top-level comment',
         lambda c: not c.is_root,
     ),
-    SolutionResponseRule(
+    AnswerResponseRule(
         'author is mod',
-        'Comment author is a mod',
-        'Comment author is not a mod',
+        '[PASS] Comment author is a mod',
+        '[FAIL] Comment author is not a mod',
         lambda c: is_mod_comment(c),
     ),
 ]
@@ -251,22 +257,21 @@ def check_rules(rules, comment):
     return all_rules_passed
 
 
-def marks_as_solved(comment):
-    """Return True if the comment meets the criteria for marking the submission
-    as solved, False otherwise.
+def awards_point(comment):
+    """Return True if the comment meets the criteria for awarding points, False otherwise.
     """
     # TODO should enforce that only one or the other can pass?
     op_rules_pass = check_rules(OP_RESPONSE_RULES, comment)
     if op_rules_pass:
-        logging.info('OP marking submission as solved')
+        logging.info('OP awarded point')
 
     mod_rules_pass = check_rules(MOD_RESPONSE_RULES, comment)
     if mod_rules_pass:
-        logging.info('Mod marking submission as solved')
+        logging.info('Mod awarded point')
 
     mod_remove_pass = check_rules(MOD_REMOVE_RULES, comment)
     if mod_remove_pass:
-        logging.info('Mod removing point')
+        logging.info('Mod removed point')
 
     return op_rules_pass or mod_rules_pass, mod_remove_pass, mod_rules_pass or mod_remove_pass
 
@@ -275,13 +280,13 @@ def is_mod_comment(comment):
     return comment.subreddit.moderator(redditor=comment.author)
 
 
-def is_valid_tag(solved_comment, valid_tags):
+def is_valid_tag(awarder_comment, valid_tags):
     """Return True if this comments post has one of the allowed tags, False otherwise.
     If there aren't any tags configured, skip this check"""
     if valid_tags is None:
         return True
 
-    submission_title = solved_comment.submission.title.lower()
+    submission_title = awarder_comment.submission.title.lower()
 
     for valid_tag in valid_tags:
         if f"[{valid_tag}]" in submission_title:
@@ -290,12 +295,12 @@ def is_valid_tag(solved_comment, valid_tags):
     return False
 
 
-def find_solver_and_comment(solved_comment):
-    """Determine the redditor responsible for solving the question."""
+def find_helper_and_comment(awarder_comment):
+    """Determine the redditor responsible for helping."""
     # TODO plz make this better someday
-    # return solved_comment.parent().author
-    solution_comment = solved_comment.parent()
-    return solution_comment.author, solution_comment
+    # return awarder_comment.parent().author
+    helper_comment = awarder_comment.parent()
+    return helper_comment.author, helper_comment
 
 
 ### Print & Logging Functions ###